requirements.txt
ipython
requests
websocket-client
